# backend/device-service/http/go/api_delete_charging_station.go
/*
 * OCPP Device Service
 *
 * REST API to manage OCPP devices (e.g. Charging Stations)
 *
 * API version: v2.0.0
 * Contact: gr.szalay@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

import (
	"fmt"
	"net/http"

	"github.com/gorilla/mux"
	"github.com/gregszalay/ocpp-csms/device-service/db"
	log "github.com/sirupsen/logrus"
)

func DeleteChargingStation(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")

	// get id from path parameter
	vars := mux.Vars(r)
	id, ok := vars["id"]
	if !ok {
		log.Error("failed to retrieve charging station id from URL path parameters")
	}

	log.Info("Attempting to delete charging station with id ", id)

	if err := db.DeleteChargingStation(id); err != nil {
		log.Error("failed to delete charging station ", id)
		w.WriteHeader(http.StatusInternalServerError)
		w.Write([]byte(fmt.Sprintf("Failed to delete charging station %s", id)))
		return
	}
	w.WriteHeader(http.StatusOK)
}